# Conflicts:
#	src/Controller/Pathfinder.java

+++ src/Model/ICollidable.java
@@ -4,8 +4,8 @@ package Model;
  * Created by Martin on 24/02/2016.
  */
 public interface ICollidable {
-	double getX();
-	double getY();
+	float getX();
+	float getY();
 	double getCollisionRadius();
 	//TODO ENUM for type
 }
